+++ src/LsUpg.hs
@@ -7,6 +7,7 @@
 ------------------------------------------------------------------------------
 
 {-# LANGUAGE LambdaCase #-}
+{-# LANGUAGE RecordWildCards #-}
 
 module LsUpg
   ( -- * Constants
@@ -54,7 +55,7 @@ import qualified Data.Yaml as Yaml
 
 -- (lsupg)
 import qualified LsUpg.Component as Component
-import LsUpg.Component (Component)
+import LsUpg.Component (Component, Options(Options, mDebugHandle, mNixPath))
 import qualified LsUpg.Component.Apk
 import qualified LsUpg.Component.Apt
 import qualified LsUpg.Component.Dnf
@@ -67,7 +68,7 @@ import qualified Paths_lsupg as Project
 ------------------------------------------------------------------------------
 -- $Constants
 
--- | lsupg version string (\"@lsupg-haskell X.X.X@\")
+-- | lsupg version string (\"@lsupg-haskell X.X.X.X@\")
 --
 -- @since 0.1.0.0
 version :: String
@@ -135,16 +136,17 @@ lookupComponent name =
 
 -- | Run specified components
 --
--- @since 0.1.0.0
+-- @since 0.3.0.0
 run
   :: [Component]
-  -> Handle        -- ^ output handle
-  -> Maybe Handle  -- ^ optional debug handle
+  -> Handle          -- ^ output handle
+  -> Maybe Handle    -- ^ optional debug handle
+  -> Maybe FilePath  -- ^ optional Nix path
   -> OutputFormat
-  -> IO Bool       -- ^ 'True' when upgrades are available
-run components outHandle mDebugHandle outputFormat = do
+  -> IO Bool         -- ^ 'True' when upgrades are available
+run components outHandle mDebugHandle mNixPath outputFormat = do
     items <- fmap concat . forM components $ \component ->
-      Component.run component mDebugHandle
+      Component.run component Options{..}
     case outputFormat of
       OutputHuman -> TLIO.hPutStr outHandle $ table
         [ [ TTC.render $ Component.componentName item
@@ -163,12 +165,13 @@ run components outHandle mDebugHandle outputFormat = do
 
 -- | Run all components
 --
--- @since 0.1.0.0
+-- @since 0.3.0.0
 runAll
-  :: Handle        -- ^ output handle
-  -> Maybe Handle  -- ^ optional debug handle
+  :: Handle          -- ^ output handle
+  -> Maybe Handle    -- ^ optional debug handle
+  -> Maybe FilePath  -- ^ optional Nix path
   -> OutputFormat
-  -> IO Bool       -- ^ 'True' when upgrades are available
+  -> IO Bool         -- ^ 'True' when upgrades are available
 runAll = run allComponents
 
 ------------------------------------------------------------------------------